Use the built-in OpenID-Connect flow of the XWiki OIDC Provider
We were using a custom flow provided by the Nextcloud extention for XWiki.
We are getting rid of it in favor of the one provided with the OIDC Provider.
At this point, the Nextcloud extension for XWiki is not technically needed
anymore. But we are keeping it since we plan to provide features on XWiki
for Nextcloud that will require a dedicated XWiki extension.